refactor python compiler
